+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 - (Feature) (Platform) Adjust Gateway timeouts
 - (Feature) (Platform) Docs and Installer improvements
 - (Feature) Promote RestartPolicy Always Feature
+- (Maintenance) Update Dependencies
+- (Feature) Backup Retry Until and BackupPolicy Until Propagation feature
 - (Maintenance) Update Envoy to v1.32.5
 
 ## [1.2.47](https://github.com/arangodb/kube-arangodb/tree/1.2.47) (2025-03-28)